requirements.txt
flask
torch
transformers
pillow
numpy
pandas
requests